refactor(ui): move UI components from ui/ to atoms/ directory
