+++ components.json
@@ -13,7 +13,7 @@
   "aliases": {
     "components": "@/components",
     "utils": "@/utils/cn",
-    "ui": "@/components/ui",
+    "ui": "@/components/atoms",
     "lib": "@/utils",
     "hooks": "@/hooks"
   },
