+++ components/Idea/More.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import { Button } from '@/components/ui/button';
+import { Button } from '@/components/atoms/button';
 
 interface MoreProps {
   isMore: boolean;
