+++ components/Milestones/MilestoneCard.tsx
@@ -14,7 +14,7 @@ import toast from 'react-hot-toast';
 import dayjs from 'dayjs';
 import { cn } from '@/utils/cn';
 import { MdSend, MdClose, MdEdit } from 'react-icons/md';
-import { Button } from '@/components/ui/button';
+import { Button } from '@/components/atoms/button';
 import DateRangePicker from '@/shared/components/DateRangePicker';
 import Form from '@/shared/components/Form';
 import { useDialog } from '@/contexts/Dialog';
