+++ layout/components/Header/MobileMenu.tsx
@@ -8,8 +8,8 @@ import {
   Collapsible,
   CollapsibleContent,
   CollapsibleTrigger,
-} from "@/components/ui/collapsible";
-import { Button } from "@/components/ui/button";
+} from "@/components/atoms/collapsible";
+import { Button } from "@/components/atoms/button";
 import { cn } from "@/utils/cn";
 
 interface OnCloseProps {
