+++ shared/components/Upload.tsx
@@ -1,7 +1,7 @@
 import toast from 'react-hot-toast';
 import { ChangeEvent, useEffect, useRef, useState } from 'react';
 import { uploadImagesSchema } from '@/services/modules/images';
-import { Button, type ButtonProps } from '@/components/ui/button';
+import { Button, type ButtonProps } from '@/components/atoms/button';
 
 export interface ImageDataType {
   id: string;
